refactor: split each State in its own file
